pathfinding feature renamed and changed some functions

+++ src/lib.rs
@@ -6,4 +6,4 @@ pub mod serialized;
 
 pub type TileId = u16;
 pub type MovementId = u8;
-pub type MapSize = u16;
+pub type MapSize = usize;
